Move applications utils to own file
